@@ -1,19 +1,24 @@
 [package]
 name = "swift-rs"
-version = "0.3.0"
+version = "1.0.0"
 description = "Call Swift from Rust with ease!"
-authors = ["you"]
+authors = ["The swift-rs contributors"]
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/Brendonovich/swift-rs"
-edition = "2018"
+edition = "2021"
 exclude=["/src-swift", "*.swift"]
 build = "src-rs/test-build.rs"
+
+# /bin/sh RUSTDOCFLAGS="--cfg docsrs" cargo +nightly doc --all-features
+[package.metadata."docs.rs"]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
 
 [lib]
 path = "src-rs/lib.rs"
 
 [dependencies]
-base64 = "0.13.0"
+base64 = "0.21.0"
 serde = { version = "1.0", features = ["derive"], optional = true}
 serde_json = { version = "1.0", optional = true }
 
@@ -25,5 +30,5 @@ serde_json = { version = "1.0" }
 serial_test = "0.10"
 
 [features]
-default = ["build"]
+default = []
 build = ["serde", "serde_json"]
