@@ -1,14 +1,14 @@
-use swift_rs::{SRObject, SRObjectArray, SRString};
+use swift_rs::{swift, Bool, Int, SRObject, SRObjectArray, SRString};
 
 #[repr(C)]
 struct Volume {
     pub name: SRString,
     path: SRString,
-    total_capacity: usize,
-    available_capacity: usize,
-    is_removable: bool,
-    is_ejectable: bool,
-    is_root_filesystem: bool,
+    total_capacity: Int,
+    available_capacity: Int,
+    is_removable: Bool,
+    is_ejectable: Bool,
+    is_root_filesystem: Bool,
 }
 
 #[repr(C)]
@@ -16,16 +16,17 @@ struct Test {
     pub null: bool,
 }
 
-extern "C" {
-    fn get_file_thumbnail_base64(path: &SRString) -> SRString;
-    fn get_mounts() -> SRObjectArray<Volume>;
-    fn return_nullable(null: bool) -> Option<SRObject<Test>>;
-}
+swift!(fn get_file_thumbnail_base64(path: &SRString) -> SRString);
+swift!(fn get_mounts() -> SRObjectArray<Volume>);
+swift!(fn return_nullable(null: Bool) -> Option<SRObject<Test>>);
 
 fn main() {
     let path = "/Users";
     let thumbnail = unsafe { get_file_thumbnail_base64(&path.into()) };
-    println!("length of base64 encoded thumbnail: {}", thumbnail.as_str().len());
+    println!(
+        "length of base64 encoded thumbnail: {}",
+        thumbnail.as_str().len()
+    );
 
     let mounts = unsafe { get_mounts() };
     println!("First Volume Name: {}", mounts[0].name);
