@@ -9,7 +9,7 @@ fn main() {
     println!("cargo:rerun-if-env-changed=TEST_SWIFT_RS");
 
     #[cfg(feature = "build")]
-    if std::env::var("TEST_SWIFT_RS").unwrap_or("false".into()) == "true" {
+    if std::env::var("TEST_SWIFT_RS").unwrap_or_else(|_| "false".into()) == "true" {
         use build::SwiftLinker;
 
         SwiftLinker::new("10.15")
