@@ -1,13 +1,44 @@
-use crate::swift;
+use crate::swift::{self, SwiftObject};
 use std::{ffi::c_void, ops::Deref, ptr::NonNull};
 
-#[repr(transparent)]
-pub struct SRObject<T>(NonNull<SRObjectImpl<T>>);
-
+#[doc(hidden)]
 #[repr(C)]
-struct SRObjectImpl<T> {
+pub struct SRObjectImpl<T> {
     _nsobject_offset: u8,
     data: T,
+}
+
+/// Wrapper for arbitrary `NSObject` types.
+///
+/// When returning an `NSObject`, its Rust type must be wrapped in `SRObject`.
+/// The type must also be annotated with `#[repr(C)]` to ensure its memory layout
+/// is identical to its Swift counterpart's.
+///
+/// ```rust
+/// use swift_rs::{swift, SRObject, Int, Bool};
+///
+/// #[repr(C)]
+/// struct CustomObject {
+///     a: Int,
+///     b: Bool
+/// }
+///
+/// swift!(fn get_custom_object() -> SRObject<CustomObject>);
+///
+/// let value = unsafe { get_custom_object() };
+///
+/// let reference: &CustomObject = value.as_ref();
+/// ```
+/// [_corresponding Swift code_]()
+#[repr(transparent)]
+pub struct SRObject<T>(pub(crate) NonNull<SRObjectImpl<T>>);
+
+impl<T> SwiftObject for SRObject<T> {
+    type Shape = T;
+
+    fn get_object(&self) -> &SRObject<Self::Shape> {
+        self
+    }
 }
 
 impl<T> Deref for SRObject<T> {
